src/facial/feature_extractor.py
import cv2
import numpy as np
from typing import Dict, List, Tuple
import dlib

class FacialFeatureExtractor:
    def __init__(self):
        # Initialize dlib's face detector and facial landmark predictor
        self.detector = dlib.get_frontal_face_detector()
        self.predictor = dlib.shape_predictor("models/shape_predictor_68_face_landmarks.dat")
        
    def extract_features(self, video_path: str) -> Dict[str, np.ndarray]:
        """
        Extract facial features from a video file.
        
        Args:
            video_path: Path to the video file
            
        Returns:
            Dictionary containing various facial features
        """
        cap = cv2.VideoCapture(video_path)
        features = []
        
        while cap.isOpened():
            ret, frame = cap.read()
            if not ret:
                break
                
            # Convert to grayscale
            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
            
            # Detect faces
            faces = self.detector(gray)
            
            for face in faces:
                # Get facial landmarks
                landmarks = self.predictor(gray, face)
                
                # Extract Action Units (simplified version)
                # In a real implementation, you would use OpenFace or similar
                # to extract actual Action Units
                features.append(self._extract_action_units(landmarks))
        
        cap.release()
        
        if not features:
            raise ValueError("No faces detected in the video")
            
        # Average features across all frames
        return {
            'action_units': np.mean(features, axis=0)
        }
    
    def _extract_action_units(self, landmarks) -> np.ndarray:
        """
        Extract simplified Action Units from facial landmarks.
        This is a simplified version - in production, use OpenFace.
        """
        # Convert landmarks to numpy array
        points = np.array([[p.x, p.y] for p in landmarks.parts()])
        
        # Calculate some basic facial measurements
        # These are simplified versions of actual Action Units
        eye_width = np.linalg.norm(points[36] - points[39])  # Left eye width
        eye_height = np.linalg.norm(points[37] - points[41])  # Left eye height
        mouth_width = np.linalg.norm(points[48] - points[54])  # Mouth width
        mouth_height = np.linalg.norm(points[51] - points[57])  # Mouth height
        
        return np.array([eye_width, eye_height, mouth_width, mouth_height])
    
    def get_feature_dimension(self) -> int:
        """Return the total dimension of all features"""
        return 4  # Number of simplified Action Units 