requirements.txt
numpy>=1.21.0
pandas>=1.3.0
librosa>=0.8.1
opencv-python>=4.5.3
tensorflow-cpu>=2.8.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
flask>=2.0.1
pytest>=6.2.5
python-dotenv>=0.19.0
dlib>=19.24.0
flake8>=4.0.1 